ScrapyNews/requirements.txt
django==2.1.5
pymysql
requests
bs4
djangorestframework
elasticsearch
scrapy
mysqlclient-1.4.6-cp37-cp37m-win_amd64.whl
scrapy_djangoitem
channels
channels_redis==2.3.3
confluent_kafka
pandas
